Revert "Budemi android main completed - Android Team"
