--- android/app/src/main/java/com/example/myapplication/model/change_password_failure_model.kt
@@ -1,8 +1,0 @@
-package com.example.myapplication.model
-
-import com.google.gson.annotations.SerializedName
-
-data class change_password_failure_model (
-    @SerializedName("detail")
-    val detail:String
-)
