--- android/app/src/main/java/com/example/myapplication/model/change_password_model.kt
@@ -1,10 +1,0 @@
-package com.example.myapplication.model
-
-import com.google.gson.annotations.SerializedName
-
-data class change_password_model (
-    @SerializedName("old_pass")
-    val old_pass:String,
-    @SerializedName("new_pass")
-    val new_pass:String
-)
