--- android/app/src/main/java/com/example/myapplication/model/change_password_response_model.kt
@@ -1,13 +1,0 @@
-package com.example.myapplication.model
-
-import com.google.gson.annotations.SerializedName
-
-class change_password_response_model (
-    @SerializedName("status")
-    val status:Integer,
-    @SerializedName("code")
-    val code:String,
-    @SerializedName("message")
-    val message:String,
-
-)
