--- android/app/src/main/java/com/example/myapplication/model/sign_in_model.kt
@@ -1,10 +1,0 @@
-package com.example.myapplication.model
-
-import com.google.gson.annotations.SerializedName
-
-data class sign_in_model(
-    @SerializedName("username")
-    val username:String,
-    @SerializedName("password")
-    val password:String
-    )
