--- android/app/src/main/java/com/example/myapplication/model/sign_up_receive_model.kt
@@ -1,10 +1,0 @@
-package com.example.myapplication.model
-
-import com.google.gson.annotations.SerializedName
-
-data class sign_up_receive_model(
-    @SerializedName("expiry")
-    val expiry:String,
-    @SerializedName("token")
-    val token:String
-)
