+++ android/app/src/main/java/com/example/myapplication/service/sign_up_api.kt
@@ -1,15 +1,16 @@
 package com.example.myapplication.service
 
 
-import com.example.myapplication.model.sign_up_send_model
-import com.example.myapplication.model.sign_up_receive_model
+import com.example.myapplication.model.sign_up_model
+import com.example.myapplication.model.sign_up_response_model
 import retrofit2.Call
 import retrofit2.http.Body
+import retrofit2.http.Headers
 import retrofit2.http.POST
 
 interface sign_up_api {
 
     //@Headers("Content-Type: application/json")
     @POST("register/")
-    fun addUser(@Body userData: sign_up_send_model): Call<sign_up_receive_model>
+    fun addUser(@Body userData: sign_up_model): Call<sign_up_response_model>
 }
