+++ android/app/src/main/java/com/example/myapplication/view/SignUpActivity.kt
@@ -3,14 +3,14 @@ package com.example.myapplication.view
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import android.widget.TextView
 import com.example.myapplication.R
-import com.example.myapplication.model.sign_up_send_model
-import com.example.myapplication.service.sign_up_api_call
+import com.example.myapplication.model.sign_up_model
+import com.example.myapplication.service.RestApiService
 import com.google.android.material.textfield.TextInputEditText
 
-var user_token=""
 
 class SignUpActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -18,13 +18,8 @@ class SignUpActivity : AppCompatActivity() {
         setContentView(R.layout.activity_sign_up)
     }
 
-    fun goToLoginPage() {
+    fun goToLoginPage(view: View) {
         var intent= Intent(applicationContext, SignInActivity::class.java)
-        startActivity(intent)
-    }
-
-    fun goToHomePage(){
-        var intent= Intent(applicationContext, HomeActivity::class.java)
         startActivity(intent)
     }
 
@@ -33,8 +28,8 @@ class SignUpActivity : AppCompatActivity() {
          val emailView=findViewById(R.id.sign_up_email) as TextInputEditText
          val passwordView=findViewById(R.id.sign_up_password) as TextInputEditText
 
-        val apiService = sign_up_api_call()
-        val userInfo = sign_up_send_model(
+        val apiService = RestApiService()
+        val userInfo = sign_up_model(
             username = userIdView.text.toString(),
             email = emailView.text.toString(),
             password = passwordView.text.toString())
@@ -44,10 +39,8 @@ class SignUpActivity : AppCompatActivity() {
             val success_message=findViewById(R.id.success_message) as TextView
             success_message.setVisibility(View.VISIBLE)
             if(it?.token!=null){
-                user_token=it?.token
-                success_message.text="Registration is successful!\n You are redirected to Homepage"
+                success_message.text="Registration is successful!"
                 success_message.postDelayed({success_message.setVisibility(View.INVISIBLE)},2000)
-                success_message.postDelayed({goToHomePage()},2000)
             }
             else{
                 success_message.text="Registration is unsuccessful!"
