--- android/app/src/main/res/layout/activity_change_password.xml
@@ -1,126 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#CFD4F1"
-        tools:context=".SignInActivity"
-        tools:layout_editor_absoluteX="16dp"
-        tools:layout_editor_absoluteY="336dp">
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/oldPasswordLayout"
-            style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-            android:layout_width="350dp"
-            android:layout_height="wrap_content"
-            app:counterMaxLength="25"
-
-            app:endIconMode="password_toggle"
-
-            app:helperText="Required*"
-
-            app:helperTextTextColor="@color/purple_700"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.491"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.453">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/oldPassword"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="New Password"
-                android:maxLength="25" />
-
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/newPasswordLayout"
-                style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-                android:layout_width="350dp"
-                android:layout_height="wrap_content"
-                app:counterMaxLength="25"
-
-                app:endIconMode="password_toggle"
-                app:helperText="Required*"
-
-                app:helperTextTextColor="@color/purple_700">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/newPassword"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:hint="Old Password"
-                    android:maxLength="25" />
-
-            </com.google.android.material.textfield.TextInputLayout>
-
-
-        </com.google.android.material.textfield.TextInputLayout>
-
-        <Button
-            android:id="@+id/button9"
-            android:layout_width="135dp"
-            android:layout_height="49dp"
-            android:layout_marginBottom="28dp"
-            android:onClick="changePassword"
-            android:text="SUBMIT"
-            app:layout_constraintBottom_toTopOf="@+id/success_message"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.273"
-            app:layout_constraintStart_toEndOf="@+id/button8" />
-
-        <Button
-            android:id="@+id/button8"
-            android:layout_width="135dp"
-            android:layout_height="49dp"
-            android:layout_marginStart="68dp"
-            android:layout_marginBottom="28dp"
-            android:onClick="goToHomePage"
-            android:text="return"
-            app:layout_constraintBottom_toTopOf="@+id/success_message"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <TextView
-            android:id="@+id/textView4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="116dp"
-            android:text="Change Password"
-            android:textColor="#6200EE"
-            android:textSize="34sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <TextView
-            android:id="@+id/success_message"
-            android:layout_width="306dp"
-            android:layout_height="90dp"
-            android:ems="10"
-            android:shadowColor="#cf1d1d"
-            android:shadowDx="0.0"
-            android:shadowDy="0.0"
-            android:shadowRadius="8"
-            android:textAlignment="center"
-            android:textColor="#cf1d1d"
-            android:textFontWeight="5"
-            android:textIsSelectable="false"
-            android:textSize="20sp"
-            android:textStyle="bold|italic"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.495"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.876" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
