+++ android/app/src/main/res/layout/activity_landing.xml
@@ -12,7 +12,7 @@
         android:layout_width="190dp"
         android:layout_height="90dp"
         android:layout_marginStart="4dp"
-        android:onClick="loggedIn"
+        android:onClick="logged_in"
         android:text="Browse"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
