+++ android/app/src/main/res/layout/activity_sign_up.xml
@@ -175,23 +175,24 @@
         <TextView
             android:id="@+id/success_message"
             android:layout_width="306dp"
-            android:layout_height="90dp"
+            android:layout_height="31dp"
             android:ems="10"
             android:shadowColor="#cf1d1d"
             android:shadowDx="0.0"
             android:shadowDy="0.0"
             android:shadowRadius="8"
-            android:textAlignment="center"
             android:textColor="#cf1d1d"
             android:textFontWeight="5"
             android:textIsSelectable="false"
             android:textSize="20sp"
+            android:textAlignment="center"
             android:textStyle="bold|italic"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.495"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/button5" />
+            app:layout_constraintTop_toBottomOf="@+id/button5"
+            app:layout_constraintVertical_bias="0.439" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
