+++ android/app/src/main/res/values/strings.xml
@@ -1,4 +1,3 @@
 <resources>
     <string name="app_name">My Application</string>
-    <string name="todo">TODO</string>
 </resources>
